+++ chess/piezas/king.py
@@ -5,10 +5,10 @@ from chess.piece import Piece
 class King(Piece):
     def __init__(self, color, board):
         super().__init__(color, board)
-        self.__symbol__ = "♚" if color == "WHITE" else "♔"
+        self.__king_symbol__ = "♚" if color == "WHITE" else "♔"
 
     def __str__(self):
-        return self.__symbol__
+        return self.__king_symbol__
 
     def possible_moves(self, row, col):
         # Movimientos posibles en las ocho direcciones
@@ -16,24 +16,11 @@ class King(Piece):
         return [
             (row + dr, col + dc)
             for dr, dc in directions
-            if self.__is__valid__move(row + dr, col + dc)
+            if self.__is_valid_move__(row + dr, col + dc)
         ]
 
-    def __is__valid__move(self, row, col):
+    def __is_valid_move__(self, row, col):
         if not (0 <= row < 8 and 0 <= col < 8):
             return False
         piece_at_position = self.__board__.get_piece(row, col)
         return piece_at_position is None or piece_at_position.__color__ != self.__color__
-
-
-
-
-
-
-
-
-
-
-
-    #def __str__(self):
-     #   return self.__white_symbol__ if self.__color__ == "WHITE" else self.__black_symbol__